Refactor toggle switch buttons code to avoid code duplication
